+++ ForJaeRin/Global/Modifier/FontModifier.swift
@@ -8,23 +8,34 @@
 import Foundation
 import SwiftUI
 
-func customFont(size: CGFloat) -> Font {
-    if let font = NSFont(name: "Pretendard-Regular", size: size) {
-        return Font(font)
-    } else {
-        // Return a fallback font or handle the error
-        return Font.system(size: size)
+struct SystemFontModifier: ViewModifier {
+    @Environment(\.sizeCategory) var sizeCategory
+    var font: SystemFont = .body
+
+    func body(content: Content) -> some View {
+        content.font(font.value)
     }
 }
 
-// MARK: 샘플코드입니다.
-
-struct CustomFontModifier: ViewModifier {
-    @Environment(\.sizeCategory) var sizeCategory
-    var name: String = "Font_Name"
-    var size: Double
-
-    func body(content: Content) -> some View {
-        return content.font(.custom(name, size: size))
+enum SystemFont {
+    case headline
+    case subTitle
+    case body
+    case caption1
+    case caption2
+    
+    var value: Font {
+        switch self {
+        case .headline:
+            return Font.systemHeadline
+        case .subTitle:
+            return Font.systemSubtitle
+        case .body:
+            return Font.systemBody
+        case .caption1:
+            return Font.systemCaption1
+        case .caption2:
+            return Font.systemCaption1
+        }
     }
 }
